test(test): add the full component matcher set
Complete the #9 matcher seed into a full, composable library: every private Matcher<Component> factory is now public, so matchers compose with Kotest's and/or and negate with shouldNot/invert alongside the shouldHave/shouldBe sugar.

Add children-order (haveChildren), exact-content (haveContent), and per-decoration sugar (shouldBeBold/Italic/Underlined/Strikethrough/Obfuscated). Split the 850-line ComponentMatchers god-file into focused per-feature files, mirror the tests, and document the catalogue in the test-module README.
