+++ modules/test/src/main/kotlin/io/github/lmliam/kotventure/test/text/SelectorMatchers.kt
@@ -1,0 +1,75 @@
+package io.github.lmliam.kotventure.test.text
+
+import io.kotest.matchers.Matcher
+import io.kotest.matchers.MatcherResult
+import io.kotest.matchers.should
+import net.kyori.adventure.text.Component
+import net.kyori.adventure.text.SelectorComponent
+
+/**
+ * Matches a selector component whose pattern is [expected]. Combine with `and`/`or` or negate with `shouldNot`.
+ */
+public fun haveSelectorPattern(expected: String): Matcher<SelectorComponent> =
+    Matcher { value ->
+        val actual = value.pattern()
+        MatcherResult(
+            actual == expected,
+            { "Expected selector pattern <$expected>, but was <$actual>." },
+            { "Expected selector pattern not to be <$expected>." },
+        )
+    }
+
+/**
+ * Matches a selector component whose separator is [expected].
+ */
+public fun haveSelectorSeparator(expected: Component): Matcher<SelectorComponent> =
+    Matcher { value ->
+        val actual = value.separator()
+        MatcherResult(
+            actual == expected,
+            { "Expected selector separator <$expected>, but was <${actual ?: "null"}>." },
+            { "Expected selector separator not to be <$expected>." },
+        )
+    }
+
+/**
+ * Matches a selector component that has no separator.
+ */
+public fun haveNoSelectorSeparator(): Matcher<SelectorComponent> =
+    Matcher { value ->
+        val actual = value.separator()
+        MatcherResult(
+            actual == null,
+            { "Expected selector separator to be absent, but was <${actual ?: "null"}>." },
+            { "Expected selector separator to be present." },
+        )
+    }
+
+/**
+ * Asserts that this component is a [SelectorComponent] and returns it typed.
+ */
+public fun Component.shouldBeSelectorComponent(): SelectorComponent = asComponentType("selector")
+
+/**
+ * Asserts that this selector component has [expected] as its selector pattern.
+ */
+public infix fun SelectorComponent.shouldHaveSelectorPattern(expected: String): SelectorComponent =
+    apply {
+        this should haveSelectorPattern(expected)
+    }
+
+/**
+ * Asserts that this selector component has [expected] as its separator.
+ */
+public infix fun SelectorComponent.shouldHaveSelectorSeparator(expected: Component): SelectorComponent =
+    apply {
+        this should haveSelectorSeparator(expected)
+    }
+
+/**
+ * Asserts that this selector component has no separator.
+ */
+public fun SelectorComponent.shouldNotHaveSelectorSeparator(): SelectorComponent =
+    apply {
+        this should haveNoSelectorSeparator()
+    }
