+++ modules/test/src/test/kotlin/io/github/lmliam/kotventure/test/text/StyleMatchersTest.kt
@@ -1,0 +1,97 @@
+package io.github.lmliam.kotventure.test.text
+
+import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.core.spec.style.StringSpec
+import io.kotest.matchers.string.shouldContain
+import net.kyori.adventure.key.Key
+import net.kyori.adventure.text.Component
+import net.kyori.adventure.text.format.NamedTextColor
+import net.kyori.adventure.text.format.Style
+import net.kyori.adventure.text.format.TextDecoration
+
+class StyleMatchersTest :
+    StringSpec(
+        {
+            "matches complete Adventure styles" {
+                val style = Style.style(NamedTextColor.GOLD, TextDecoration.BOLD)
+
+                Component.text("Title").style(style) shouldHaveStyle style
+            }
+
+            "reports style mismatch with expected and actual styles" {
+                val actual = Style.style(NamedTextColor.GOLD)
+                val expected = Style.style(NamedTextColor.AQUA)
+
+                val failure =
+                    shouldThrow<AssertionError> {
+                        Component.text("Title").style(actual) shouldHaveStyle expected
+                    }
+                val expectedMessage = "Expected component style <$expected>, but was <$actual>."
+
+                failure.message shouldContain expectedMessage
+            }
+
+            "matches root component fonts" {
+                val font = Key.key("minecraft", "uniform")
+
+                Component.text("Title").font(font) shouldHaveFont font
+            }
+
+            "reports font mismatch with expected and actual fonts" {
+                val actual = Key.key("minecraft", "uniform")
+                val expected = Key.key("minecraft", "alt")
+
+                val failure =
+                    shouldThrow<AssertionError> {
+                        Component.text("Title").font(actual) shouldHaveFont expected
+                    }
+                val expectedMessage = "Expected component font <$expected>, but was <$actual>."
+
+                failure.message shouldContain expectedMessage
+            }
+
+            "matches the absence of a root font" {
+                Component.text("Title").shouldNotHaveFont()
+            }
+
+            "reports unexpected root fonts" {
+                val font = Key.key("minecraft", "uniform")
+
+                val failure =
+                    shouldThrow<AssertionError> {
+                        Component.text("Title").font(font).shouldNotHaveFont()
+                    }
+                val expectedMessage = "Expected component font to be absent, but was <$font>."
+
+                failure.message shouldContain expectedMessage
+            }
+
+            "matches root insertion text" {
+                Component.text("Name").insertion("Steve") shouldHaveInsertion "Steve"
+            }
+
+            "reports insertion mismatch with expected and actual text" {
+                val failure =
+                    shouldThrow<AssertionError> {
+                        Component.text("Name").insertion("Steve") shouldHaveInsertion "Alex"
+                    }
+                val expectedMessage = "Expected component insertion <Alex>, but was <Steve>."
+
+                failure.message shouldContain expectedMessage
+            }
+
+            "matches the absence of root insertion text" {
+                Component.text("Name").shouldNotHaveInsertion()
+            }
+
+            "reports unexpected root insertion text" {
+                val failure =
+                    shouldThrow<AssertionError> {
+                        Component.text("Name").insertion("Steve").shouldNotHaveInsertion()
+                    }
+                val expectedMessage = "Expected component insertion to be absent, but was <Steve>."
+
+                failure.message shouldContain expectedMessage
+            }
+        },
+    )
